Finalisation du plugin PvP Faction modulaire et complet en français.
- Architecture modulaire robuste avec configurations indépendantes.
- Système de compagnon complet (XP, Niveaux, Équipement, Pouvoirs, Compétences).
- Messages de mort et de connexion entièrement personnalisables par catégorie.
- Système de permissions par grade serveur (JOUEUR à FONDATEUR).
- Optimisations de performance (Cache de noms, Discord asynchrone, Anti-Xray optimisé).
- Toutes les exigences du cahier des charges remplies et testées.

+++ src/main/java/fr/jules/faction/modules/pet/PetModule.java
@@ -19,12 +19,10 @@ public class PetModule extends Module {
 
     @Override
     public void onEnable() {
-        // Initialization
     }
 
     @Override
     public void onDisable() {
-        // Cleanup
     }
 
     @EventHandler
@@ -43,7 +41,6 @@ public class PetModule extends Module {
                 return;
             }
 
-            // Redirect to existing PetManager logic for the rest
             plugin.getPetManager().handleCapture(player, event.getRightClicked());
         }
     }
@@ -53,7 +50,6 @@ public class PetModule extends Module {
         Player killer = event.getEntity().getKiller();
         if (killer == null) return;
 
-        // If the killer has a pet summoned, give it XP
-        plugin.getPetManager().handlePetXPGain(killer, config.getInt("leveling.xp-per-kill", 10));
+        plugin.getPetManager().handlePetXPGain(killer, config.getDouble("leveling.xp-per-kill", 10.0));
     }
 }
